Themes controller: move logged-in handlers to a new module not imported by SSR
