+++ client/webpack.config.node.js
@@ -165,10 +165,6 @@ const webpackConfig = {
 			COMMIT_SHA: JSON.stringify( commitSha ),
 			'process.env.NODE_ENV': JSON.stringify( bundleEnv ),
 		} ),
-		new webpack.NormalModuleReplacementPlugin(
-			/^calypso[/\\]my-sites[/\\]themes[/\\]theme-upload$/,
-			'calypso/components/empty-component'
-		), // Depends on BOM
 		new webpack.IgnorePlugin( { resourceRegExp: /^\.\/locale$/, contextRegExp: /moment$/ } ),
 		! isDevelopment && new ExternalModulesWriter(),
 	].filter( Boolean ),
